requirements.txt
beautifulsoup4==4.12.3
cssutils==2.11.1
curl_cffi==0.7.1
pillow==10.3.0
pytesseract==0.3.10
python-decouple==3.8
python-telegram-bot==21.4
requests==2.32.2
retry==0.9.2
supabase==2.6.0
